refactor(CLIENT): :fire: Refactor the files structure
and add sounds when you click on the login
