+++ EnglishApp/app/src/assets/adaptive-icon.png
