+++ EnglishApp/app/src/assets/logos/logo2.png
